Enhancement/8/add functions to add discount and task as percentage customer phone from contacts and ui corrections

+++ lib/Api/BusinessPage/update_draft.dart
@@ -766,8 +766,8 @@ class _UpdateDraftState extends State<UpdateDraft> {
               });
         });
   }
-  int calculateTotal() {
-    int total = TotalAndSubTotal().getSubTotal(context, isNewInvoice);
+  double calculateTotal() {
+    double total = TotalAndSubTotal().getSubTotal(context, isNewInvoice);
 
     if(_children.length == 0){
